allure_pytest_selenium/requirements.txt
pytest
selenium
pytest-selenium
pytest-allure
page-objects
